refactor(models): extract CanvasConfig trait into its own module
