+++ src/dto_models/config.rs
@@ -6,6 +6,7 @@
 mod appendix_config;
 mod book_config;
 mod book_layout_solver_config;
+mod canvas_config;
 mod cover_config;
 mod fitness_weights;
 mod page_layout_solver_config;
@@ -13,8 +14,9 @@ mod preview_config;
 mod project_config;
 
 pub use appendix_config::AppendixConfig;
-pub use book_config::{BookConfig, CanvasConfig};
+pub use book_config::BookConfig;
 pub use book_layout_solver_config::{BookLayoutSolverConfig, ValidationError};
+pub use canvas_config::CanvasConfig;
 pub use cover_config::{CoverConfig, CoverMode, SpineConfig};
 pub use fitness_weights::FitnessWeights;
 pub use page_layout_solver_config::PageLayoutSolverConfig;
